Need to sort the derivedStates on each column so they don't mutate the list on position change.
I think I'll do the changing with the modal by updating the composition local from the ModalService. That way, the composition in view in the bg will not have a working knob listener.

The knob listener doesn't know what's calling it, so there's no way to know inside inputEvents.collect {} when to ignore events or not, so we need to be more heavy-handed with swapping out the CompositionLocal.