Fix Two column so that list is recomposed on modal popup.
I think what was causing it was the old list undernearth was remembered {} which was great for keeping it in composition, but it had a down-stream effect that the old scroll listener attached to it was also around.

Now I invalidate the remember { circularList } when the compositionlocal (lol, should be a composition global) changes with the knob scroll listener.

The next step is to figure out how to store the list that goes out of composition. I want the previously isSelected to stick around, but no knob scroll on it.